Укажи export и import для модулей
